+++ architect_py/async_client.py
@@ -44,7 +44,7 @@ from architect_py.grpc.orderflow import OrderflowChannel
 from architect_py.grpc.resolve_endpoint import PAPER_GRPC_PORT, resolve_endpoint
 from architect_py.utils.nearest_tick import TickRoundMethod
 from architect_py.utils.orderbook import update_orderbook_side
-from architect_py.utils.pandas import candles_to_dataframe
+from architect_py.utils.pandas import candles_to_dataframe, tickers_to_dataframe
 from architect_py.utils.price_bands import price_band_pairs
 from architect_py.utils.symbol_parsing import nominative_expiration
 
@@ -806,24 +806,16 @@ class AsyncClient:
 
         """
         grpc_client = await self.hmart()
-        if start.tzinfo is not timezone.utc:
-            raise ValueError(
-                "start must be a utc datetime:\n"
-                "for example datetime.now(timezone.utc) or \n"
-                "dt = datetime(2025, 4, 15, 12, 0, 0, tzinfo=timezone.utc)"
-            )
-        if end.tzinfo is not timezone.utc:
-            raise ValueError(
-                "end must be a utc datetime:\n"
-                "for example datetime.now(timezone.utc) or \n"
-                "dt = datetime(2025, 4, 15, 12, 0, 0, tzinfo=timezone.utc)"
-            )
+        if not start.tzinfo:
+            raise ValueError("start time must be timezone-aware")
+        if not end.tzinfo:
+            raise ValueError("end time must be timezone-aware")
         req = HistoricalCandlesRequest(
             venue=venue,
             symbol=str(symbol),
             candle_width=candle_width,
-            start_date=start,
-            end_date=end,
+            start_date=start.astimezone(timezone.utc),
+            end_date=end.astimezone(timezone.utc),
         )
         res: HistoricalCandlesResponse = await grpc_client.unary_unary(req)
 
@@ -889,6 +881,34 @@ class AsyncClient:
         req = TickerRequest(symbol=str(symbol), venue=venue)
         res: Ticker = await grpc_client.unary_unary(req)
         return res
+
+    async def get_tickers(
+        self,
+        *,
+        venue: Venue,
+        symbols: Optional[Sequence[TradableProduct | str]] = None,
+        include_options: bool = False,
+        sort_by: Optional[SortTickersBy | str] = None,
+        offset: Optional[int] = None,
+        limit: Optional[int] = None,
+        as_dataframe: bool = False,
+    ) -> Union[Sequence[Ticker], pd.DataFrame]:
+        grpc_client = await self.marketdata(venue)
+        sort_by = SortTickersBy(sort_by) if sort_by else None
+        symbols = [str(symbol) for symbol in symbols] if symbols else None
+        req = TickersRequest.new(
+            offset=offset,
+            include_options=include_options,
+            sort_by=sort_by,
+            limit=limit,
+            symbols=symbols,
+            venue=venue,
+        )
+        res: TickersResponse = await grpc_client.unary_unary(req)
+        if as_dataframe:
+            return tickers_to_dataframe(res.tickers)
+        else:
+            return res.tickers
 
     async def stream_l1_book_snapshots(
         self,